Move core functions to separate module
It can be okay to have functions defined in `__ini__`, but it generally
isn't a good pattern. For example, pylint won't pick up on unused imports,
etc. etc.

This commit moves the function definitions in `__init__` to a separate
`core` module. This causes pylint to start complaining, and this will be
addressed subsequently.
